telecom/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
boto3==1.26.0
botocore==1.29.0
distro==1.7.0
dj-database-url==1.2.0
Django==3.2.23
django-allauth==0.54.0
django-debug-toolbar==3.8.1
django-storages==1.13.2
gunicorn==20.1.0
huggingface-hub==0.16.4
jiter==0.6.0
jmespath==0.10.0
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.23.4
python-dotenv==1.0.0
s3transfer==0.6.0
sentry-sdk==1.29.2
sqlparse==0.4.4
whitenoise==6.5.0
supabase==1.0.3
